+++ test/pay/stripe/payment_method_test.rb
@@ -96,4 +96,18 @@ class Pay::Stripe::PaymentMethodTest < ActiveSupport::TestCase
     refute pm2.default?
     assert pm3.default?
   end
+
+  test "Stripe sync records the customer's stripe_account when none is given" do
+    @pay_customer.update!(stripe_account: "acct_123")
+    ::Stripe::Customer.stubs(:retrieve).returns(::Stripe::Customer.construct_from(invoice_settings: {default_payment_method: nil}))
+    pay_payment_method = Pay::Stripe::PaymentMethod.sync("pm_123", object: fake_stripe_payment_method)
+    assert_equal "acct_123", pay_payment_method.stripe_account
+  end
+
+  test "Stripe sync_payment_intent passes the stripe_account through" do
+    payment_intent = ::Stripe::PaymentIntent.construct_from(id: "pi_123", object: "payment_intent", payment_method: fake_stripe_payment_method)
+    ::Stripe::PaymentIntent.expects(:retrieve).with({id: "pi_123", expand: ["payment_method"]}, {stripe_account: "acct_123"}).returns(payment_intent)
+    Pay::Stripe::PaymentMethod.expects(:sync).with("pm_123", object: payment_intent.payment_method, stripe_account: "acct_123")
+    Pay::Stripe::PaymentMethod.sync_payment_intent("pi_123", stripe_account: "acct_123")
+  end
 end